refactored constants and configurationmanager
